test_requirements.txt
black==19.10b0
coverage==5.2
codecov==2.1.7
devtools==0.5.1
factory-boy==2.12.0
isort==5.1.3
flake8==3.8.3
pycodestyle==2.6.0
pyflakes==2.2.0
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-sugar==0.9.4
pytest-xdist==1.33.0
